Updated music hack and added potential stable bundle in apps.
- Removed v1 Advanced music hack for v2.1 with NESC support (3kb)
- Added Option Menu v1.1 NESC/SNESC support (61kb)
- Added RetroArch 'Neo' 1.7.1d, the latest stable release of RA intended as the default go to 'stable' release. Includes fixes for NESC/SNESC (2.21mb)

These are purely optional. it just bumps the download up 2mb but with the recent savings in the executable icon I believe this is a viable option. With the above additions the users effectively have a "starter" pack to be getting on with out the box.

All downloads have been tested and vetted and release to public. SCC is available and not considered as 'nightly' builds.