Update SnesGame.cs
Hakchi2 CE will remove (external) header or any SNES rom, independently of extension. It is important because sometimes a header is added manually by the user in order for him to be able to apply a patch that requires a headered rom. Said situation may result in a headered .sfc rom, and Hakchi2 CE currently will not remove such header (the header would be removed only if the extension were .smc).